Fix: Mapped Girls, FollowUps and Deliveries Code refactor - GetData - Data Formatters
